requirements.txt
baidu-aip==2.2.17.0
boto==2.49.0
boto3==1.9.196
botocore==1.12.196
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Session==0.3.1
gensim==3.8.0
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lxml==4.3.4
MarkupSafe==1.1.1
numpy==1.16.4
pycparser==2.19
PyExecJS==1.5.1
pymongo==3.8.0
python-dateutil==2.8.0
redis==3.2.1
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.0
six==1.12.0
smart-open==1.8.4
urllib3==1.25.3
Werkzeug==0.15.4
